+++ src/materialkatalogview.cpp
@@ -95,7 +95,7 @@ void MaterialKatalogView::slNewTemplate()
 
   if( parentItem ) {
     // try to find out which catalog is open/current
-      newMat->setChapter( getKatalog( MaterialCatalogName )->chapterID( parentItem->text( 0 ) ) );
+      newMat->setChapter( getKatalog( MaterialCatalogName )->chapterID( parentItem->text( 0 ) ).toInt() );
   }
 
   mNewItem = matListView->addMaterialToView( parentItem, newMat );
